// 2025 territoriale/secret-storage/website/db.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface Entries {
  created_at: Generated<Date>;
  id: Generated<number>;
  name: string;
  owner: number;
  secret: string;
}

export interface Users {
  id: Generated<number>;
  username: string;
}

export interface DB {
  entries: Entries;
  users: Users;
}